test: migrate zero-slack-browser-connect to bdd
